+++ fiddles/vue/fiddle-0000-TemplateJs/src/main.js
@@ -5,6 +5,7 @@ import App from './App.vue'
 import router from './router'
 
 import './assets/main.css'
+import 'bootswatch/dist/sketchy/bootstrap.css'
 
 const app = createApp(App)