+++ fiddles/vue/template/js/src/router/index.js
@@ -1,23 +1,43 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
+
+
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
-      name: 'home',
-      component: HomeView
+      name: 'fiddle',
+      component: () => import('../views/FiddleView.vue')
+    },
+    {
+      path: '/docs',
+      name: 'docs',
+      component: () => import('../views/HomeView.vue')
     },
     {
       path: '/about',
       name: 'about',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
       component: () => import('../views/AboutView.vue')
-    }
-  ]
+    },
+    {
+      path: '/404',
+      component: () => import('../views/404View.vue')
+    },
+    {
+      path: '/:catchAll(.*)',
+      beforeEnter: onBeforeError,
+      redirect: '/404'
+    },
+    ]
 })
+
+function onBeforeError(to, from, next) {
+  if (from.fullPath === '/' || from.fullPath.includes('index.html')) {
+    next('/')
+  } else {
+    next()
+  }
+}
 
 export default router