+++ scripts/bin/vue/_create.sh
@@ -57,11 +57,22 @@ function vueCreateJs() {
             rm -r .gitignore;
         fi
         
-        $(voidSubstr '{{FiddleName}}' ${fiddle} "package.json";) || exit 5;
         $(voidSubstr '{{FiddleName}}' ${fiddle} "index.html";) || exit 5;
         $(voidSubstr '{{Author}}' ${AUTHOR} "index.html";) || exit 5;
+
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "vite.config.js";) || exit 5;
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "package.json";) || exit 5;
+
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "src/App.vue";) || exit 5;
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "src/components/TopNav.vue";) || exit 5;
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "src/router/index.js";) || exit 5;
+        
         $(voidSubstr '{{FiddleName}}' ${fiddle} "README.md";) || exit 5;
         $(voidSubstr '{{BornOnDate}}' ${bornOnDate} "README.md";) || exit 5;
+
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "src/assets/README.md";) || exit 5;
+        $(voidSubstr '{{BornOnDate}}' ${bornOnDate} "src/assets/README.md";) || exit 5;
+ 
        
 
         npmInstallDep || exit $?;
@@ -123,6 +134,7 @@ function vueCreateTs() {
         $(voidSubstr '{{Author}}' ${AUTHOR} "index.html";) || exit 5;
 
         $(voidSubstr '{{FiddleName}}' ${fiddle} "vite.config.ts";) || exit 5;
+        $(voidSubstr '{{FiddleName}}' ${fiddle} "package.json";) || exit 5;
         $(voidSubstr '{{FiddleName}}' ${fiddle} "src/App.vue";) || exit 5;
         $(voidSubstr '{{FiddleName}}' ${fiddle} "src/components/TopNav.vue";) || exit 5;
         $(voidSubstr '{{FiddleName}}' ${fiddle} "src/router/index.ts";) || exit 5;